add more logs
added request logger
and sentry TracingInterceptors for auction controller
moved "logging" functions to dir
