+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "unique-marketplace-api",
-  "version": "1.0.26",
+  "version": "1.0.27",
   "description": "Backend project for unique marketplace",
   "author": "Unique Network",
   "private": true,
